readme, more refactoring
