+++ react-client/src/components/overlay/OverlayBanner.jsx
@@ -7,11 +7,7 @@ import {
   BannerWrapper,
 } from '../styled-components/BannerStyles.jsx';
 
-import {
-  ButtonLeft,
-  ButtonRight,
-} from '../styled-components/ButtonStyles.jsx';
-
+import OverlayNav from './OverlayNavButtons.jsx'
 
 const BannerSpacer = styled.div`
   margin-top: 10px;
@@ -25,17 +21,7 @@ export default props => {
         {props.render()}
         {props.children}
         {props.displayCarouselButtons ? (
-          <div>
-            <ButtonLeft
-              onClick={props.handleLeftClick}
-              src='https://cdn0.iconfinder.com/data/icons/basic-ui-elements-round/700/01_arrow_left-128.png'
-            />
-
-            <ButtonRight
-              onClick={props.handleRightClick}
-              src='https://cdn0.iconfinder.com/data/icons/basic-ui-elements-round/700/01_arrow_left-128.png'
-            />
-          </div>
+          <OverlayNav handleLeftClick={props.handleLeftClick} handleRightClick={props.handleRightClick}/>
         ) : null}
       </BannerWrapper>
       <BannerSpacer />
