+++ react-client/src/components/styled-components/NotFoundStyles.jsx
